Fixed SSE overflow

+++ Source/API/EbSvtAv1.h
@@ -101,9 +101,9 @@ typedef struct EbBufferHeaderType {
     // pic info
     uint32_t qp;
     uint32_t pic_type;
-    uint32_t luma_sse;
-    uint32_t cr_sse;
-    uint32_t cb_sse;
+    uint64_t luma_sse;
+    uint64_t cr_sse;
+    uint64_t cb_sse;
     // pic flags
     uint32_t flags;
